+++ Changes.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+  - Adding an optional timeout to `await_read` and `await_write`.
+  - Allow fibers to return complex data (AV*, HV*).
+
 ## [v0.0.2] - 2026-02-17
 
 ### Fixed
@@ -13,7 +19,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
   - Made unit tests a lot more noisy
-
 
 ## [v0.0.1] - 2026-02-16